requirements.txt
# Data Science & Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Geospatial
geopandas>=0.14.0
rasterio>=1.3.0
shapely>=2.0.0
pyproj>=3.6.0

# Google Earth Engine
earthengine-api>=0.1.370

# HTTP / APIs
requests>=2.31.0
openmeteo-requests>=0.3.0
requests-cache>=1.2.0
retry-requests>=2.0.0

# Computer Vision / Deep Learning
torch>=2.0.0, <=2.8.1
torchvision>=0.15.0, <=0.23.1
transformers>=4.38.0
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
jupyterlab>=4.0.0
